Correcting syntax in anasyn.py + printing function in idtable.py
